Upgrade to sbt 2
sbt-houserules 0.12.0 ships an _sbt2_3 build, which removed the last
blocker to moving this plugin to sbt 2. sbt 1 support is dropped: the
modules now build with Scala 3.8.4 and publish as
sbt-blowout-{core,json-circe,yaml-circe}_sbt2_3.

Three sbt 2 changes needed real work beyond the version bumps:

Bare settings are injected into every subproject in sbt 2, not just the
root. Left as they were, `noPublishSettings` would have set
`publish / skip` on all three modules and `name := "sbt-blowout"` would
have given them the same artifact id. They now live in an explicit root
project that aggregates the modules.

Tasks assigned with `:=` are cached by default, so sbt would return a
cached value and skip the file system effects that blowoutGenerate and
blowoutCheck exist for. Both are wrapped in `Def.uncached`, and a new
`generate-is-not-cached` scripted test guards against a regression.

The meta build resolved sbt-blowout-yaml-circe 0.2.1 to generate this
project's own workflows, but an sbt 2 meta build can only load _sbt2_3
plugins and the first such release is the one this build produces. It
now compiles modules/core and modules/yaml-circe from source, so no
published artifact is needed and blowoutCheck always validates HEAD.

Scala 3 details: `Def.Setting[_]` becomes `[?]`, `url` is deprecated in
favour of `uri`, the `blowout` package object becomes top level
definitions, and `import sbt.*` still needs the `io` exclusion because
`sbt.io` shadows the root `io` package. `val Version = new { ... }`
widens to Object under Scala 3 and an `object` in build.sbt is invisible
to later statements, so the version list moved to project/Version.scala.

Also drops the now redundant `OrganizeImports.targetDialect` override
(houserules defaults to Auto) and bumps actions/checkout to v7 and
actions/setup-java to v5, whose v4 is deprecated.
