+++ modules/core/src/main/scala/io/taig/blowout/BlowoutComment.scala
@@ -1,0 +1,13 @@
+package io.taig.blowout
+
+type BlowoutComment = List[String] => Option[String]
+
+object BlowoutComment {
+  val Disabled: BlowoutComment = _ => None
+
+  def prefixed(prefix: String): BlowoutComment = lines =>
+    Some(lines.map(line => s"$prefix$line").mkString("", "\n", "\n"))
+
+  val Hash: BlowoutComment = prefixed("# ")
+  val Slash: BlowoutComment = prefixed("// ")
+}
