+++ modules/core/src/main/scala/io/taig/blowout/BlowoutPlugin.scala
@@ -3,12 +3,12 @@ package io.taig.blowout
 import com.github.difflib.DiffUtils
 import com.github.difflib.UnifiedDiffUtils
 import sbt.Keys.state
-import sbt.io.IO
 import sbt.plugins.JvmPlugin
-import sbt.{io => _, _}
+// `sbt.io` would shadow the root `io` package that this plugin lives in
+import sbt.{io as _, *}
 
 import java.io.FileNotFoundException
-import scala.jdk.CollectionConverters._
+import scala.jdk.CollectionConverters.*
 
 object BlowoutPlugin extends AutoPlugin {
   object autoImport extends BlowoutKeys {
@@ -19,13 +19,13 @@ object BlowoutPlugin extends AutoPlugin {
     val BlowoutComment = io.taig.blowout.BlowoutComment
   }
 
-  import autoImport._
+  import autoImport.*
 
   override def requires: Plugins = JvmPlugin
 
   override def trigger: PluginTrigger = noTrigger
 
-  override def projectSettings: Seq[Def.Setting[_]] = Def.settings(
+  override def projectSettings: Seq[Def.Setting[?]] = Def.settings(
     blowoutGenerators := Nil,
     blowoutHeader := List(
       "This file was automatically generated by sbt-blowout and should not be edited manually.",
@@ -34,13 +34,14 @@ object BlowoutPlugin extends AutoPlugin {
     blowoutRender := { generator =>
       Some(blowoutHeader.value).filter(_.nonEmpty).flatMap(generator.comment).getOrElse("") + generator.content()
     },
-    blowoutGenerate := {
+    // sbt 2 caches tasks by default, which would skip the file system effects these tasks exist for
+    blowoutGenerate := Def.uncached {
       blowoutGenerators.value.map { generator =>
         IO.write(generator.target, blowoutRender.value(generator))
         generator.target
       }
     },
-    blowoutCheck := {
+    blowoutCheck := Def.uncached {
       val log = state.value.log
 
       blowoutGenerators.value.foreach { generator =>
