+++ modules/core/src/sbt-test/blowout/check-on-different-content/build.sbt
@@ -1,4 +1,4 @@
-TaskKey[Unit]("alterGeneratedFile") := {
+TaskKey[Unit]("alterGeneratedFile") := Def.uncached {
   IO.write(file("foobar.yml"), "fizbuz")
 }
 
@@ -6,4 +6,4 @@ enablePlugins(BlowoutPlugin)
 
 blowoutGenerators += BlowoutGenerator.strict(file("foobar.yml"), content = "foobar")
 
-scalaVersion := "3.1.0"
+scalaVersion := "3.8.4"
