+++ modules/core/src/sbt-test/blowout/generate-is-not-cached/build.sbt
