+++ modules/core/src/sbt-test/blowout/generate-is-not-cached/project/plugins.sbt
@@ -1,0 +1,5 @@
+sys.props.get("plugin.version") match {
+  case Some(version) => addSbtPlugin("io.taig" % "sbt-blowout-core" % version)
+  case _ => sys.error("""|The system property 'plugin.version' is not defined.
+                         |Specify this property using the scriptedLaunchOpts -D.""".stripMargin)
+}
