+++ modules/core/src/sbt-test/blowout/generator-not-run/build.sbt
@@ -2,4 +2,4 @@ enablePlugins(BlowoutPlugin)
 
 blowoutGenerators += BlowoutGenerator.strict(file("foobar.yml"), content = "foobar")
 
-scalaVersion := "3.1.0"
+scalaVersion := "3.8.4"
