+++ project/Version.scala
@@ -1,0 +1,7 @@
+object Version {
+  val Circe = "0.14.15"
+  val CirceYaml = "0.15.2"
+  val Java = "17"
+  val JavaDiffUtils = "4.17"
+  val Scala = "3.8.4"
+}
