+++ project/build.properties
@@ -1,1 +1,1 @@
-sbt.version=1.12.12
+sbt.version=2.0.6
